[with examples] - Fix siconos script to allow a proper handling of addons when running outside of the source dir - Fix numerics warnings (int<->size_t conversions ...)

+++ numerics/src/tools/JordanAlgebra.h
@@ -39,9 +39,8 @@ NumericsMatrix* Arrow_repr(const double* const vec, size_t vecSize, size_t varsC
  * This is an alternative of Arrow_repr, without creation of matrix
  * (not allocation of memory)
  */
-void Arrow_repr_replace(NumericsMatrix* Arw_mat, const double* const vec,
-                      const unsigned int vecSize, const size_t varsCount);
-
+void Arrow_repr_replace(NumericsMatrix* Arw_mat, const double* const vec, const size_t vecSize,
+                        const size_t varsCount);
 
 /**
  * Returns reflection matrix.
